% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{binary-string-concat}
\alias{\%p\%}
\alias{\%s\%}
\alias{binary-string-concat}
\title{Concatenate two strings.}
\usage{
x \%p\% y

x \%s\% y
}
\arguments{
\item{x, y}{A character vector}
}
\description{
\code{\%p\%} and \code{\%s\%} are wrappers for \code{paste0(..., collapse = '')} and
\code{paste0(..., collapse = ' ')}, respectively, that combine two character vectors.
}
\examples{
'the quick brown fox jum' \%p\% 'ped over the lazy dog'

gen_sql <- function(column, table) "SELECT" \%s\% column \%s\% "FROM" \%s\% table

}

