% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{contat-two-strings}
\alias{\%p\%}
\alias{\%s\%}
\alias{contat-two-strings}
\title{Concatenate two strings.}
\usage{
x %p% y
x %s% y

x \%s\% y
}
\arguments{
\item{x}{A character vector}

\item{y}{A character vector}
}
\description{
`%p%`and `%s%` are wrappers for `paste0(..., collapse = '')` and
`paste0(..., collapse = ' ')`, respectively, that combine two character vectors.
}
\examples{
'the quick brown fox jum' \%p\% 'ped over the lazy dog'

gen_sql <- function(column, table) "SELECT" \%s\% column \%s\% "FROM" \%s\% table
}

