% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format-string}
\alias{\%f\%}
\alias{format-string}
\title{Pass variables into strings}
\usage{
x %f% y
}
\arguments{
\item{x}{A character vector}

\item{y}{A (possibly named) character vector}
}
\description{
Pass variables into strings using pairs of curly brackets
to identify points of insertion.
}
\examples{
# order matters when not using a named vector
'the quick {} fox jumped {} the lazy {}' \%f\% c('brown', 'over', 'dog')

gen_sql_query <- function(column, table, id){
    query <- "SELECT {col} FROM {tab} WHERE pk = {id}"
    query \%f\% c('col' = column, 'tab' = table, 'id' = id)
}

gen_sql_query('LASTNAME', 'STUDENTS', '12345')

}

